update tree nodes
